Reworked wave scaling to use integers instead of doubles Fixed incorrect wave counts which previously caused waves to end early

+++ src/condor/event/waves/AllFlamingWave.java
@@ -8,10 +8,10 @@ import condor.phantom.PhantomType;
 
 public class AllFlamingWave extends Wave {
 
-  private static TreeMap<PhantomType, Double> map = new TreeMap<>();
+  private static TreeMap<PhantomType, Integer> map = new TreeMap<>();
 
   static {
-    map.put(PhantomType.FLAMING_PHANTOM, 6.0);
+    map.put(PhantomType.FLAMING_PHANTOM, 6);
   }
 
   public AllFlamingWave() {
